+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 * interpretation examples
 * possibility to write `C%` instead of `C*%`
 * slightly change colour depending on the plot function expression, so that we have a chance to distinguish different lines plotting the same territory
+* scatter chart: NC/C